Feature/eint23 450 validate contract proposal

See merge request intellecteu/blockchain/clients/equilend-integration/eint-1source-integration!91

+++ src/main/resources/db/migration/V1__initial_db.sql
@@ -253,7 +253,7 @@ CREATE TABLE IF NOT EXISTS trade
     venue_id        BIGINT       NULL,
     instrument_id   BIGINT       NULL,
     rate_id         BIGINT       NULL,
-    quantity        INT          NULL,
+    quantity        DOUBLE PRECISION    NULL,
     currency        VARCHAR(255) NULL,
     dividend_rate   DOUBLE PRECISION NULL,
     trade_date      DATE         NULL,
